v1.5.0
-- Fix for "Stim Rejection not working for auto/manual time exclusion"

-- Fixed the issue of propagation of excluded channel indices from first wavelength to the second. Added sanity check/test to confirm it is working.

-- Critical bug fix SplineSG mlAct
SplineSG has mlActAuto as an input. Due to the changes in homer3, mlActAuto containes only the first wavelength. SplineSG thus failed to clean the data from the second wavelength.
Note for homer2 users: This was working properly for homer2.
Note for homer3 users: Please reconsider running your pipeline with this version and up.

-- Add new function hmrR_TaskNorm.m to calculate the average of the HRF over the duration of the task.